+++ contracts/IexecInterfaceToken.sol
@@ -1,54 +1,42 @@
+// SPDX-FileCopyrightText: 2020-2025 IEXEC BLOCKCHAIN TECH <contact@iex.ec>
 // SPDX-License-Identifier: Apache-2.0
-
-/******************************************************************************
- * Copyright 2020 IEXEC BLOCKCHAIN TECH                                       *
- *                                                                            *
- * Licensed under the Apache License, Version 2.0 (the "License");            *
- * you may not use this file except in compliance with the License.           *
- * You may obtain a copy of the License at                                    *
- *                                                                            *
- *     http://www.apache.org/licenses/LICENSE-2.0                             *
- *                                                                            *
- * Unless required by applicable law or agreed to in writing, software        *
- * distributed under the License is distributed on an "AS IS" BASIS,          *
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
- * See the License for the specific language governing permissions and        *
- * limitations under the License.                                             *
- ******************************************************************************/
 
 pragma solidity ^0.6.0;
 pragma experimental ABIEncoderV2;
 
-import "./modules/interfaces/IOwnable.sol";
-import "./modules/interfaces/IexecAccessors.sol";
-import "./modules/interfaces/IexecCategoryManager.sol";
-import "./modules/interfaces/IexecERC20.sol";
-import "./modules/interfaces/IexecEscrowToken.sol";
-import "./modules/interfaces/IexecEscrowTokenSwap.sol";
-import "./modules/interfaces/IexecMaintenance.sol";
-import "./modules/interfaces/IexecOrderManagement.sol";
-import "./modules/interfaces/IexecPoco1.sol";
-import "./modules/interfaces/IexecPoco2.sol";
-import "./modules/interfaces/IexecRelay.sol";
-import "./modules/interfaces/IexecTokenSpender.sol";
-import "./modules/interfaces/ENSIntegration.sol";
+import "./interfaces/IOwnable.sol";
+import "./interfaces/IexecAccessors.sol";
+import "./interfaces/IexecCategoryManager.sol";
+import "./interfaces/IexecERC20.sol";
+import "./interfaces/IexecEscrowToken.sol";
+import "./interfaces/IexecEscrowTokenSwap.sol";
+import "./interfaces/IexecConfiguration.sol";
+import "./interfaces/IexecOrderManagement.sol";
+import "./interfaces/IexecPoco1.sol";
+import "./interfaces/IexecPoco2.sol";
+import "./interfaces/IexecRelay.sol";
+import "./interfaces/IexecTokenSpender.sol";
 
-
+/**
+ * A global interface that aggregates all the interfaces needed to interact with
+ * the PoCo contracts in token mode.
+ * @dev Referenced in the SDK with the current path `contracts/IexecInterfaceToken.sol`.
+ * Changing the name or the path would cause a breaking change in the SDK.
+ */
 interface IexecInterfaceToken is
-	IOwnable,
-	IexecAccessors,
-	IexecCategoryManager,
-	IexecERC20,
-	IexecEscrowToken,
-	IexecEscrowTokenSwap,
-	IexecMaintenance,
-	IexecOrderManagement,
-	IexecPoco1,
-	IexecPoco2,
-	IexecRelay,
-	IexecTokenSpender,
-	ENSIntegration
+    IOwnable,
+    IexecAccessors,
+    IexecCategoryManager,
+    IexecERC20,
+    IexecEscrowToken,
+    IexecEscrowTokenSwap,
+    IexecConfiguration,
+    IexecOrderManagement,
+    IexecPoco1,
+    IexecPoco2,
+    IexecRelay,
+    IexecTokenSpender
 {
-	receive()  external override(IexecEscrowToken, IexecEscrowTokenSwap) payable;
-	fallback() external override(IexecEscrowToken, IexecEscrowTokenSwap) payable;
+    receive() external payable override(IexecEscrowToken, IexecEscrowTokenSwap);
+    fallback() external payable override(IexecEscrowToken, IexecEscrowTokenSwap);
 }