+++ contracts/tools/testing/slither/Slither.sol
@@ -3,16 +3,16 @@
 
 pragma solidity ^0.8.0;
 
-import {IexecPocoAccessorsDelegate} from "../../../modules/delegates/IexecPocoAccessorsDelegate.sol";
-import {IexecPoco1Delegate} from "../../../modules/delegates/IexecPoco1Delegate.sol";
-import {IexecPoco2Delegate} from "../../../modules/delegates/IexecPoco2Delegate.sol";
+import {IexecPocoAccessorsFacet} from "../../../facets/IexecPocoAccessorsFacet.sol";
+import {IexecPoco1Facet} from "../../../facets/IexecPoco1Facet.sol";
+import {IexecPoco2Facet} from "../../../facets/IexecPoco2Facet.sol";
 
 /**
  * @notice This contract is dedicated to slither analysis.
  *
- * @dev This contract aggregates multiple delegate contracts into a single entry point
+ * @dev This contract aggregates multiple facet contracts into a single entry point
  *      to facilitate static analysis using Slither.
  */
 //slither-disable-start unused-state
-contract Slither is IexecPocoAccessorsDelegate, IexecPoco1Delegate, IexecPoco2Delegate {}
+contract Slither is IexecPocoAccessorsFacet, IexecPoco1Facet, IexecPoco2Facet {}
 //slither-disable-end unused-state