+++ scripts/set-callback-gas.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { deployments, ethers } from 'hardhat';
-import { IexecAccessors__factory, IexecMaintenanceDelegate__factory } from '../typechain';
+import { IexecAccessors__factory, IexecConfigurationFacet__factory } from '../typechain';
 
 (async () => {
     const requestedCallbackGas = Number(process.env.CALLBACK_GAS);
@@ -12,13 +12,13 @@ import { IexecAccessors__factory, IexecMaintenanceDelegate__factory } from '../t
     }
     console.log(`Setting callback-gas to ${requestedCallbackGas.toLocaleString()} ..`);
     const [owner] = await ethers.getSigners();
-    const erc1538ProxyAddress = (await deployments.get('ERC1538Proxy')).address;
+    const diamondProxyAddress = (await deployments.get('Diamond')).address;
     const viewCallbackGas = async () =>
-        (await IexecAccessors__factory.connect(erc1538ProxyAddress, owner).callbackgas())
+        (await IexecAccessors__factory.connect(diamondProxyAddress, owner).callbackgas())
             .toNumber()
             .toLocaleString();
     const callbackGasBefore = await viewCallbackGas();
-    await IexecMaintenanceDelegate__factory.connect(erc1538ProxyAddress, owner)
+    await IexecConfigurationFacet__factory.connect(diamondProxyAddress, owner)
         .setCallbackGas(requestedCallbackGas)
         .then((tx) => tx.wait());
     console.log(`Changed callback-gas from ${callbackGasBefore} to ${await viewCallbackGas()}`);